students/IS22/Azimov Adam/laba_3/requirements.txt
pymilvus
sentence-transformers
numpy
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymilvus==2.3.7
python-dotenv==1.0.0
numpy==1.26.4
setuptools==80.0.0  # Фиксируем версию, чтобы избежать предупреждения